fix(codex): tighten plugin thread config
