+++ extensions/codex/src/app-server/plugin-thread-config.ts
@@ -3,13 +3,9 @@
  * for native Codex turns.
  */
 import crypto from "node:crypto";
-import { embeddedAgentLog } from "openclaw/plugin-sdk/agent-harness-runtime";
 import {
   defaultCodexAppInventoryCache,
-  serializeCodexAppInventoryError,
-  type CodexAppInventorySnapshot,
   type CodexAppInventoryCache,
-  type CodexAppInventoryRequest,
 } from "./app-inventory-cache.js";
 import {
   resolveCodexPluginsPolicy,
@@ -25,13 +21,23 @@ import {
   readCodexPluginInventory,
   type CodexPluginInventory,
   type CodexPluginInventoryDiagnostic,
-  type CodexPluginInventoryRecord,
   type CodexPluginOwnedApp,
   type CodexPluginRuntimeRequest,
 } from "./plugin-inventory.js";
 import type { CodexPluginMetadataCache } from "./plugin-metadata-cache.js";
-import type { CodexAppServerRequestResult } from "./protocol.js";
-import { isJsonObject, type JsonObject, type JsonValue, type v2 } from "./protocol.js";
+import {
+  collectInventoryOwnedAppIds,
+  readConfigLayersForAppAdmission,
+  readThreadAdmissibleAccountApps,
+  refreshAppInventoryNow,
+  resolveExplicitAppEnablement,
+  resolvePluginAppThreadAdmission,
+  resolveThreadConfigAppsForRecord,
+  shouldForceRefreshForNotReadyPluginApps,
+  toOwnedAccountApp,
+  type CodexPluginThreadAppAdmissionDiagnostic,
+} from "./plugin-thread-app-admission.js";
+import { isJsonObject, type JsonObject, type JsonValue } from "./protocol.js";
 
 /** Policy context for one app id exposed by a configured Codex plugin. */
 export type PluginAppPolicyContextEntry = {
@@ -66,12 +72,12 @@ export type PluginAppPolicyContext = {
 /** Diagnostic emitted while building app config for a native Codex thread. */
 type CodexPluginThreadConfigDiagnostic =
   | CodexPluginInventoryDiagnostic
+  | CodexPluginThreadAppAdmissionDiagnostic
   | {
       code:
         | "plugin_activation_failed"
         | "plugin_config_timeout"
         | "app_not_ready"
-        | "account_app_inventory_unavailable"
         | "approval_overrides_clear_failed";
       plugin?: ResolvedCodexPluginPolicy;
       message: string;
@@ -625,199 +631,12 @@ function shouldRefreshMissingAppInventory(
   );
 }
 
-async function refreshAppInventoryNow(
-  params: BuildCodexPluginThreadConfigParams,
-  appCache: CodexAppInventoryCache,
-  options: { forceRefetch?: boolean; reason?: string; targetAppIds?: readonly string[] } = {},
-): Promise<CodexAppInventorySnapshot | undefined> {
-  const appCacheKey = params.appCacheKey;
-  if (!appCacheKey) {
-    return undefined;
-  }
-  const request: CodexAppInventoryRequest = async (method, requestParams) =>
-    (await params.request(method, requestParams)) as CodexAppServerRequestResult<typeof method>;
-  try {
-    const snapshot = await appCache.refreshNow({
-      key: appCacheKey,
-      request,
-      nowMs: params.nowMs,
-      forceRefetch: options.forceRefetch,
-      targetAppIds: options.targetAppIds,
-    });
-    return snapshot;
-  } catch (error) {
-    embeddedAgentLog.warn("codex plugin thread config app inventory refresh failed", {
-      reason: options.reason,
-      forceRefetch: options.forceRefetch === true,
-      error: serializeCodexAppInventoryError(error),
-    });
-    // Keep building from the diagnostic inventory state; app exposure remains scoped below.
-    return undefined;
-  }
-}
-
-function collectInventoryOwnedAppIds(inventory: CodexPluginInventory): string[] {
-  return Array.from(
-    new Set(inventory.records.flatMap((record) => record.ownedAppIds).filter(Boolean)),
-  ).toSorted();
-}
-
 function emptyCodexPluginInventory(policy: ResolvedCodexPluginsPolicy): CodexPluginInventory {
   return {
     policy,
     records: [],
     diagnostics: [],
   };
-}
-
-async function readThreadAdmissibleAccountApps(
-  params: BuildCodexPluginThreadConfigParams,
-  appCache: CodexAppInventoryCache,
-): Promise<{
-  apps: v2.AppInfo[];
-  source?: CodexAppInventorySnapshot["source"];
-  diagnostic?: CodexPluginThreadConfigDiagnostic;
-}> {
-  // Account-wide mode needs metadata for every installed app, not only the
-  // configured plugin-owned app ids used by targeted startup refreshes.
-  const snapshot = await refreshAppInventoryNow(params, appCache, {
-    forceRefetch: false,
-    reason: "account_apps_all",
-    targetAppIds: [],
-  });
-  if (!snapshot) {
-    return {
-      apps: [],
-      diagnostic: {
-        code: "account_app_inventory_unavailable",
-        message: "Codex account app inventory was unavailable; account apps were not exposed.",
-      },
-    };
-  }
-  return {
-    apps: snapshot.apps
-      // Account-wide discovery must preserve Codex's effective enablement policy.
-      // Only an explicitly configured plugin may provisionally enable its owned app.
-      .filter((app) => resolveAppInfoThreadAdmission(app, snapshot.source) === "ready")
-      .toSorted((left, right) => left.id.localeCompare(right.id)),
-    source: snapshot.source,
-  };
-}
-
-function toOwnedAccountApp(app: v2.AppInfo): CodexPluginOwnedApp {
-  return {
-    id: app.id,
-    name: app.name,
-    accessible: app.isAccessible,
-    enabled: app.isEnabled,
-    needsAuth: !app.isAccessible,
-  };
-}
-
-function resolveThreadConfigAppsForRecord(params: {
-  record: CodexPluginInventoryRecord;
-  inventory: CodexPluginInventory;
-}): CodexPluginOwnedApp[] {
-  if (params.inventory.appInventory?.state === "missing") {
-    return [];
-  }
-  return params.record.apps;
-}
-
-type CodexPluginAppThreadAdmission = "ready" | "provisional" | "blocked";
-
-function resolvePluginAppThreadAdmission(
-  app: CodexPluginOwnedApp,
-  inventory: CodexPluginInventory,
-): CodexPluginAppThreadAdmission {
-  if (app.accessible) {
-    return "ready";
-  }
-  const snapshot = inventory.appInventory?.snapshot;
-  const appInfo = snapshot?.apps.find((candidate) => candidate.id === app.id);
-  return appInfo ? resolveAppInfoThreadAdmission(appInfo, snapshot.source) : "blocked";
-}
-
-function resolveAppInfoThreadAdmission(
-  app: v2.AppInfo,
-  source: CodexAppInventorySnapshot["source"] | undefined,
-): CodexPluginAppThreadAdmission {
-  if (app.isAccessible) {
-    return "ready";
-  }
-  // The explicit OpenClaw plugin entry is the unmanaged enablement decision.
-  // Codex still applies feature/workspace gates and managed app requirements
-  // after this thread override; thread-scoped attestation fails closed if they win.
-  return source === "installed" && !app.isEnabled ? "provisional" : "blocked";
-}
-
-async function readConfigLayersForAppAdmission(
-  params: BuildCodexPluginThreadConfigParams,
-): Promise<readonly JsonObject[] | undefined> {
-  try {
-    const response = await params.request("config/read", {
-      includeLayers: true,
-      ...(params.configCwd ? { cwd: params.configCwd } : {}),
-    });
-    if (!isJsonObject(response) || !Array.isArray(response.layers)) {
-      throw new Error("Codex config/read omitted config layers");
-    }
-    return response.layers.flatMap((layer) => {
-      if (!isJsonObject(layer)) {
-        throw new Error("Codex config/read returned an invalid config layer");
-      }
-      if (layer.disabledReason !== undefined && layer.disabledReason !== null) {
-        if (typeof layer.disabledReason !== "string") {
-          throw new Error("Codex config/read returned an invalid disabled layer");
-        }
-        return [];
-      }
-      if (!isJsonObject(layer.config)) {
-        throw new Error("Codex config/read returned an invalid layer config");
-      }
-      return [layer.config];
-    });
-  } catch (error) {
-    embeddedAgentLog.warn("codex plugin app admission config read failed", { error });
-    return undefined;
-  }
-}
-
-function resolveExplicitAppEnablement(
-  layersHighestPrecedenceFirst: readonly JsonObject[],
-  appId: string,
-): boolean | undefined {
-  // Codex includes disabled layers and orders active config from highest to lowest.
-  // The first app-specific value wins; ignoring `_default` preserves the deny-all
-  // service-account baseline while preventing SessionFlags from undoing an explicit opt-out.
-  for (const layer of layersHighestPrecedenceFirst) {
-    const apps = layer.apps;
-    const app = isJsonObject(apps) ? apps[appId] : undefined;
-    if (!isJsonObject(app) || !Object.hasOwn(app, "enabled")) {
-      continue;
-    }
-    return app.enabled === true;
-  }
-  return undefined;
-}
-
-function shouldForceRefreshForNotReadyPluginApps(
-  params: BuildCodexPluginThreadConfigParams,
-  policy: ResolvedCodexPluginsPolicy,
-  inventory: CodexPluginInventory,
-): boolean {
-  if (!params.appCacheKey || !policy.pluginPolicies.some((plugin) => plugin.enabled)) {
-    return false;
-  }
-  if (inventory.appInventory?.state === "missing") {
-    return false;
-  }
-  return inventory.records.some(
-    (record) =>
-      record.appOwnership === "proven" &&
-      record.ownedAppIds.length > 0 &&
-      (record.apps.length === 0 || record.apps.some((app) => !app.accessible)),
-  );
 }
 
 function policyFingerprint(policy: ResolvedCodexPluginsPolicy): JsonValue {
